chore(Refactor tests): Add test in separate files for different items e.g meals
[Delivers #157038506]
